% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_star.R
\name{d_star}
\alias{d_star}
\title{Determine the Number of Principal Components to Reach a Given Variance Threshold}
\usage{
d_star(C, rho = 0.9)
}
\arguments{
\item{C}{Numeric matrix. Covariance matrix whose eigenvalues are analyzed.}

\item{rho}{Numeric between 0 and 1. Desired proportion of total variance to explain. Default is 0.9.}
}
\value{
Integer. The minimum number of components needed to reach at least \code{rho} proportion of total variance.
}
\description{
This function calculates the minimum number of principal components required
to explain at least a specified proportion of the total variance of a covariance matrix.
}
\examples{
C <- matrix(c(2, 1, 1, 2), 2, 2)
d_star(C, rho = 0.9)

}
